+++ SeleniumJO/src/test/java/Testrail_listeners/UseAsTestRailId.java
@@ -1,4 +1,4 @@
-package TestrailAPI;
+package Testrail_listeners;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -6,7 +6,9 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD) //on method level
-public @interface TestRails {
 
-	String id() default "none";
+public @interface UseAsTestRailId {
+
+	int testRailId() default 0;
+	String[] tags() default "";
 }
