+++ SeleniumJO/test-output/emailable-report.html
@@ -8,15 +8,11 @@
 <body>
 <table>
 <tr><th>Test</th><th># Passed</th><th># Skipped</th><th># Retried</th><th># Failed</th><th>Time (ms)</th><th>Included Groups</th><th>Excluded Groups</th></tr>
-<tr><th colspan="8">User acceptance TestSuite</th></tr>
-<tr><td><a href="#t0">Expense Tracker Validattion</a></td><td class="num">1</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">12,166</td><td></td><td></td></tr>
+<tr><th colspan="8">jdohnnyjangjang</th></tr>
+<tr><td><a href="#t0">Test</a></td><td class="num">1</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">3,774</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">User acceptance TestSuite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Expense Tracker Validattion &#8212; failed (configuration methods)</th></tr><tr class="failedeven"><td rowspan="1">TestrailAPI._0_TestRail_API_Send</td><td><a href="#m0">afterTest</a></td><td rowspan="1">1676571224128</td><td rowspan="1">-1676571224128</td></tr><tr><th colspan="4">Expense Tracker Validattion &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">Script._3_Contract</td><td><a href="#m1">Contract</a></td><td rowspan="1">1676571215339</td><td rowspan="1">8572</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">jdohnnyjangjang</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="1">Script._1_Login</td><td><a href="#m0">login</a></td><td rowspan="1">1676810866536</td><td rowspan="1">430</td></tr></tbody>
 </table>
-<h2>Expense Tracker Validattion</h2><h3 id="m0">TestrailAPI._0_TestRail_API_Send#afterTest</h3><table class="result"><tr><th>Exception</th></tr><tr><td><div class="stacktrace">org.testng.TestNGException: 
-Can inject only one of &lt;ITestContext, XmlTest&gt; into a @AfterTest annotated afterTest.
-For more information on native dependency injection please refer to https://testng.org/doc/documentation-main.html#native-dependency-injection
-... Removed 21 stack frames</div></td></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m1">Script._3_Contract#Contract</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Test</h2><h3 id="m0">Script._1_Login#login</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
