+++ SeleniumJO/test-output/old/Default suite/methods.html
@@ -1,6 +1,24 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="91ef70">  <td>23/02/16 22:20:20</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_3_ContractSetupStartModalPopup.rdr3()[pri:0, instance:Script._3_ContractSetupStartModalPopup@44821a96]">rdr3</td> 
-  <td>main@1325056130</td>   <td></td> </tr>
+<tr bgcolor="f3b89c">  <td>23/02/17 22:48:10</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;_0_BaseTest.initializeWebDriver()[pri:0, instance:Script._1_Login@71def8f8]">&gt;&gt;initializeWebDriver</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="91ef70">  <td>23/02/17 22:48:13</td>   <td>3270</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_1_Login.login()[pri:0, instance:Script._1_Login@71def8f8]">login</td> 
+  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="afe297">  <td>23/02/17 22:48:17</td>   <td>6795</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_2_MainScreen.EnterContractSetupStart()[pri:0, instance:Script._2_MainScreen@73e22a3d]">EnterContractSetupStart</td> 
+  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="f3b89c">  <td>23/02/17 22:48:18</td>   <td>8344</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;_0_BaseTest.initializeWebDriver()[pri:0, instance:Script._3_Contract@47987356]">&gt;&gt;initializeWebDriver</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="92be74">  <td>23/02/17 22:48:21</td>   <td>10972</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="_3_Contract.Contract()[pri:0, instance:Script._3_Contract@47987356]">Contract</td> 
+  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="f3b89c">  <td>23/02/17 22:48:14</td>   <td>3668</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;_0_BaseTest.tearDown()[pri:0, instance:Script._1_Login@71def8f8]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="f3b89c">  <td>23/02/17 22:48:29</td>   <td>19435</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;_0_BaseTest.tearDown()[pri:0, instance:Script._3_Contract@47987356]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="f3b89c">  <td>23/02/17 22:48:14</td>   <td>3875</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;_0_BaseTest.initializeWebDriver()[pri:0, instance:Script._2_MainScreen@73e22a3d]">&gt;&gt;initializeWebDriver</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="f3b89c">  <td>23/02/17 22:48:18</td>   <td>8183</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;_0_BaseTest.tearDown()[pri:0, instance:Script._2_MainScreen@73e22a3d]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@2108763062</td>   <td></td> </tr>
+<tr bgcolor="948bbd">  <td>23/02/17 22:48:09</td>   <td>-1355</td> <td title="&gt;&gt;_0_TestRail_API_Send.createSuite(org.testng.ITestContext)[pri:0, instance:Script._0_TestRail_API_Send@7205765b]">&gt;&gt;createSuite</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@2108763062</td>   <td></td> </tr>
 </table>
