+++ SeleniumJO/test-output/old/Default suite/toc.html
@@ -9,8 +9,8 @@
 <table border='1' width='100%'>
 <tr valign='top'>
 <td>1 test</td>
-<td><a target='mainFrame' href='classes.html'>1 class</a></td>
-<td>1 method:<br/>
+<td><a target='mainFrame' href='classes.html'>3 classes</a></td>
+<td>3 methods:<br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods.html'>chronological</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-alphabetical.html'>alphabetical</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-not-run.html'>not run (0)</a></td>
@@ -20,9 +20,9 @@
 <td><a target='mainFrame' href='reporter-output.html'>reporter output</a></td>
 <td><a target='mainFrame' href='testng.xml.html'>testng.xml</a></td>
 </tr></table>
-<table width='100%' class='test-failed'>
+<table width='100%' class='test-passed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>Default test (0/1/0)</td><td valign='top' align='right'>
+<table style='width: 100%'><tr><td valign='top'>Default test (3/0/0)</td><td valign='top' align='right'>
   <a href='Default test.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
