+++ SeleniumJO/test-output/old/index.html
@@ -5,5 +5,5 @@
 <h2><p align='center'>Test results</p></h2>
 <table border='1' width='100%' class='main-page'><tr><th>Suite</th><th>Passed</th><th>Failed</th><th>Skipped</th><th>testng.xml</th></tr>
 <tr align='center' class='invocation-passed'><td><em>Total</em></td><td><em>1</em></td><td><em>0</em></td><td><em>0</em></td><td>&nbsp;</td></tr>
-<tr align='center' class='invocation-passed'><td><a href='User acceptance TestSuite/index.html'>User acceptance TestSuite</a></td>
-<td>1</td><td>0</td><td>0</td><td><a href='User acceptance TestSuite/testng.xml.html'>Link</a></td></tr></table></body></html>
+<tr align='center' class='invocation-passed'><td><a href='jdohnnyjangjang/index.html'>jdohnnyjangjang</a></td>
+<td>1</td><td>0</td><td>0</td><td><a href='jdohnnyjangjang/testng.xml.html'>Link</a></td></tr></table></body></html>
