+++ SeleniumJO/test-output/testng-failed.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Failed suite [User acceptance TestSuite]" guice-stage="DEVELOPMENT">
-  <test thread-count="5" name="Expense Tracker Validattion(failed)">
+<suite name="Failed suite [jdohnnyjangjang]" guice-stage="DEVELOPMENT">
+  <listeners>
+    <listener class-name="Testrail_listeners.TestListener"/>
+  </listeners>
+  <test thread-count="5" name="Test(failed)">
     <classes>
-      <class name="Script._3_Contract">
+      <class name="Script._1_Login">
         <methods>
-          <include name="tearDown"/>
-          <include name="beforeTest"/>
-          <include name="Contract"/>
-          <include name="initializeWebDriver"/>
+          <include name="login"/>
         </methods>
-      </class> <!-- Script._3_Contract -->
+      </class> <!-- Script._1_Login -->
     </classes>
-  </test> <!-- Expense Tracker Validattion(failed) -->
-</suite> <!-- Failed suite [User acceptance TestSuite] -->
+  </test> <!-- Test(failed) -->
+</suite> <!-- Failed suite [jdohnnyjangjang] -->
